requirements.txt
APScheduler==3.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
hse-ruz==2.1.2
idna==3.3
pyTelegramBotAPI==4.4.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
schedule==1.1.0
six==1.16.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
